utils/estree-parser: split the file
